# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API para la gestión de contactos",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                   "example": "Peter"
                },
                "lastName": {
                  "example": "Parker"
                },
                "email": {
                  "example": "pp.spider@gmail.com"
                },
                "favoriteColor": {
                   "example": "Red"
                },
                "birthday": {
                 "example": "1962-10-08"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "Peter"
                },
                "lastName": {
                  "example": "Parker"
                },
                "email": {
                  "example": "pp.spider@gmail.com"
                },
                "favoriteColor": {
                  "example": "Red"
                },
                "birthday": {
                  "example": "1962-10-08"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Smith"
        },
        "email": {
          "type": "string",
          "example": "john.smith@test.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-05-15"
        }
      }
    },
    "ContactResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "507f1f77bcf86cd799439011"
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Smith"
        },
        "email": {
          "type": "string",
          "example": "john.smith@test.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-05-15"
        }
      }
    }
  }
}